requirements.txt
# Course Generator v5/v6 Dependencies

# LLM & Embeddings
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4

# Document Processing
python-docx>=0.8.11
PyPDF2>=3.0.0

# Core
numpy>=1.24.0
sympy>=1.12
pandas>=2.0.0

# Async & HTTP
aiohttp>=3.8.0
requests>=2.31.0

# Utilities
python-dateutil>=2.8.0
